Refactor proxy fetch logic outside of mix tasks.
